+++ src/main/java/com/jmj/qnasite/controller/ArticleController.java
@@ -2,11 +2,11 @@ package com.jmj.qnasite.controller;
 
 import com.jmj.qnasite.dto.ArticleDto;
 import com.jmj.qnasite.dto.CommentDto;
-import com.jmj.qnasite.entity.Article;
-import com.jmj.qnasite.repository.ArticleRepository;
+import com.jmj.qnasite.domain.article.Article;
+import com.jmj.qnasite.domain.article.ArticleRepository;
 import com.jmj.qnasite.service.CommentService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,17 +16,14 @@ import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.util.List;
 
-@Controller
 @Slf4j
+@RequiredArgsConstructor
+@Controller
 public class ArticleController {
 
-    // refactoring - 생성자 주입 방법 변경
-    @Autowired
-    private CommentService commentService;
+    private final CommentService commentService;
 
-    // refactoring - 생성자 주입 방법 변경
-    @Autowired
-    private ArticleRepository articleRepository;
+    private final ArticleRepository articleRepository;
 
     @GetMapping("/articles/new")
     public String newArticleForm(){
