+++ src/main/java/com/jmj/qnasite/domain/article/Article.java
@@ -1,4 +1,4 @@
-package com.jmj.qnasite.entity;
+package com.jmj.qnasite.domain.article;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
