+++ src/main/resources/templates/articles/new.mustache
@@ -7,12 +7,6 @@
             <input type="text" class="form-control" id="title" name="title">
             <!--name에는 dto에 있는 필드명과 똑같이 -->
         </div>
-
-    <!--    <div class=mb-3">-->
-    <!--        <label class="form-label" for="author">작성자</label>-->
-    <!--        <input type="text" class="form-control" id="author" name="author">-->
-    <!--    </div>-->
-
         <div class="mb-3">
             <label class="form-label" for="content">내용</label>
             <textarea class="form-control" rows="3" id="content"  name="content"></textarea>
